requirements.txt
djangorestframework
psycopg2-binary
faker
requests